remove copy module
